docs: update imageRegistry context fields
Updates the documentation for the imageRegistry context source
to use the correct field names 'manifestList' and 'configData'
instead of the outdated 'imageIndex' and 'config'.

Clarifies the description of the manifestList field and its
relationship to image indexes and multi-architecture images.

Refines the description mapping context keys to crane commands
for accuracy and conciseness.

Fixes #1523

Signed-off-by: Ashish Thirunagari <53890434+ash01ish@users.noreply.github.com>