Add Kyverno and CNCF Code of Conduct to website (#1717)
* Add Reo.dev script to documentation site footer.html

This PR updates the Kyverno documentation website by adding JavaScript code generated by Reo.Dev to the footer of the website.

This script tracks and analyzes basic de-identified user interactions on the site and should not impact site speed times.

This change was suggested in direct discussions with Kyverno folks.

Reo.dev purges PII, complies with the GDPR, and Kyverno users can easily opt out of analytics in various ways documented here: https://docs.reo.dev/guide/reo-dev-javascript-cookies-and-consent-guide.

Signed-off-by: Cortney <cortney.nickerson@nirmata.com>

* Add code-of-conduct.md

This PR is to add the Kyverno and CNCF Code of Conduct to the user website for reference.

Signed-off-by: Cortney <cortney.nickerson@nirmata.com>

---------

Signed-off-by: Cortney <cortney.nickerson@nirmata.com>
Signed-off-by: kyverno-bot <104836976+kyverno-bot@users.noreply.github.com>